refactor: Cleaned up player rendering class hierarchy
